+++ examples/timestamping/backend/src/schema.rs
@@ -17,13 +17,15 @@
 use chrono::{DateTime, Utc};
 use exonum::crypto::Hash;
 use exonum_merkledb::{Entry, IndexAccess, ObjectHash, ProofMapIndex};
-use exonum_proto_derive::protobuf_convert;
+use exonum_proto_derive::ProtobufConvert;
 
 use crate::{proto, transactions::Config};
 
 /// Stores content's hash and some metadata about it.
+#[derive(
+    Serialize, Deserialize, Clone, Debug, PartialEq, ProtobufConvert, BinaryValue, ObjectHash,
+)]
 #[protobuf_convert(source = "proto::Timestamp")]
-#[derive(Serialize, Deserialize, Clone, Debug, PartialEq, BinaryValue, ObjectHash)]
 pub struct Timestamp {
     /// Hash of the content.
     pub content_hash: Hash,
@@ -43,7 +45,7 @@ impl Timestamp {
 
 /// Timestamp entry.
 #[protobuf_convert(source = "proto::TimestampEntry", serde_pb_convert)]
-#[derive(Clone, Debug, BinaryValue, ObjectHash)]
+#[derive(Clone, Debug, ProtobufConvert, BinaryValue, ObjectHash)]
 pub struct TimestampEntry {
     /// Timestamp data.
     pub timestamp: Timestamp,