+++ exonum/tests/websocket/blockchain/mod.rs
@@ -25,7 +25,7 @@ use exonum::{
     },
 };
 use exonum_merkledb::{Snapshot, TemporaryDB};
-use exonum_proto_derive::protobuf_convert;
+use exonum_proto_derive::ProtobufConvert;
 
 use std::{
     net::SocketAddr,
@@ -36,8 +36,8 @@ mod proto;
 
 pub const SERVICE_ID: InstanceId = 118;
 
+#[derive(Serialize, Deserialize, Clone, Debug, ProtobufConvert, BinaryValue, ObjectHash)]
 #[protobuf_convert(source = "proto::CreateWallet")]
-#[derive(Serialize, Deserialize, Clone, Debug, BinaryValue, ObjectHash)]
 pub struct CreateWallet {
     pub pubkey: PublicKey,
     pub name: String,
@@ -52,8 +52,8 @@ impl CreateWallet {
     }
 }
 
+#[derive(Serialize, Deserialize, Clone, Debug, ProtobufConvert, BinaryValue, ObjectHash)]
 #[protobuf_convert(source = "proto::Transfer")]
-#[derive(Serialize, Deserialize, Clone, Debug, BinaryValue, ObjectHash)]
 pub struct Transfer {
     pub from: PublicKey,
     pub to: PublicKey,