+++ testkit/examples/timestamping/main.rs
@@ -29,15 +29,15 @@ use exonum::{
     },
 };
 use exonum_merkledb::{ObjectHash, Snapshot};
-use exonum_proto_derive::protobuf_convert;
+use exonum_proto_derive::ProtobufConvert;
 use exonum_testkit::{ApiKind, InstanceCollection, TestKitBuilder};
 
 mod proto;
 
 // Simple service implementation.
 
+#[derive(Serialize, Deserialize, Clone, Debug, ProtobufConvert, BinaryValue, ObjectHash)]
 #[protobuf_convert(source = "proto::TxTimestamp")]
-#[derive(Serialize, Deserialize, Clone, Debug, BinaryValue, ObjectHash)]
 struct TxTimestamp {
     message: String,
 }